+++ tests/API/Payment/MyFatoorahPaymentTest.php
@@ -6,13 +6,14 @@ use MyFatoorah\Library\API\Payment\MyFatoorahPayment;
 
 class MyFatoorahPaymentTest extends \PHPUnit\Framework\TestCase
 {
+
     private $keys;
 
     //-----------------------------------------------------------------------------------------------------------------------------------------
     public function __construct()
     {
         parent::__construct();
-        $this->keys = include(__DIR__ . '/../../apiKeys.php');
+        $this->keys = include __DIR__ . '/../../apiKeys.php';
     }
 
     //-----------------------------------------------------------------------------------------------------------------------------------------