+++ tests/apiKeys.php
@@ -30,7 +30,7 @@ return [
         'countryCode'               => 'KWT',
         'isTest'                    => true,
         'message'                   => 'Not Authorized',
-    //        'exception'                 => 'Exception: You are not Authorized to use API.',
+        //        'exception'                 => 'Exception: You are not Authorized to use API.',
         'exception'                 => 'Invalid Account Information!, Please check the Dhl account information and try again.',
         'getCurrencyRatesException' => 'You are not Authorized to use API.',
         'refundException'           => 'Key: No transaction exist matching this Key!',